# Refactor implementation of column setitem