--- a/python/cudf/cudf/core/column/column.py
+++ b/python/cudf/cudf/core/column/column.py
@@ -599,19 +599,23 @@ class ColumnBase(Column, Serializable):
             )
         else:
             try:
+                tmp = self.as_frame()
                 if not isinstance(key, Column):
                     key = as_column(key)
                 if is_scalar(value):
-                    input = self
-                    out = input.as_frame()._scatter(key, value)._as_column()
+                    result = tmp.__class__._from_data(
+                        *libcudf.copying.scatter(value, key, tmp)
+                    )
+                    result._copy_type_metadata(tmp)
+                    out = result._as_column()
                 else:
                     if not isinstance(value, Column):
                         value = as_column(value)
-                    out = (
-                        self.as_frame()
-                        ._scatter(key, value.as_frame())
-                        ._as_column()
+                    result = tmp.__class__._from_data(
+                        *libcudf.copying.scatter(value.as_frame(), key, tmp)
                     )
+                    result._copy_type_metadata(tmp)
+                    out = result._as_column()
             except RuntimeError as e:
                 if "out of bounds" in str(e):
                     raise IndexError(