@vyasr 2021-08-24
Although not necessarily the main purpose of this PR, it's worth noting that it does improve performance a bit, although nothing groundbreaking.

Before:
```
(rapids) rapids@compose:~/cudf$ ipython -i -c "import cudf; import numpy as np; df = cudf.DataFrame({'a': 
Python 3.7.10 | packaged by conda-forge | (default, Feb 19 2021, 16:07:37)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.26.0 -- An enhanced Interactive Python. Type '?' for help.np.arange(10000)});"

In [1]: %timeit df[df > 5000] = 10
1.24 ms ± 5.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: %timeit df[df > 5000] = 10
1.25 ms ± 12.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [3]: %timeit df[df > 5000] = 10
1.24 ms ± 7.92 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

After:
```
(rapids) rapids@compose:~/cudf$ ipython -i -c "import cudf; import numpy as np; df = cudf.DataFrame({'a': np.arange(10000)});"
Python 3.7.10 | packaged by conda-forge | (default, Feb 19 2021, 16:07:37)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.26.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: %timeit df[df > 5000] = 10
1.15 ms ± 15.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: %timeit df[df > 5000] = 10
1.17 ms ± 7.39 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [3]: %timeit df[df > 5000] = 10
1.16 ms ± 3.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```